refactor: part(1) folder structure
