+++ src/api/AccommodationOffers.ts
@@ -5,8 +5,8 @@ import {
 } from '@windingtree/glider-types/dist/win';
 import { winClient } from 'src/api/winClient';
 import { daysBetween } from 'src/utils/date';
-import { SearchPropsType } from '.';
-import { getPassengersBody } from '../useAccommodationsAndOffers/helpers';
+import { SearchPropsType } from '../hooks/useAccommodation';
+import { getPassengersBody } from '../utils/getPassengerBody';
 
 export interface AccommodationResponseType {
   accommodation: WinAccommodation;
