+++ src/containers/search/Results.tsx
@@ -7,7 +7,7 @@ import { styled } from '@mui/system';
 import { daysBetween } from '../../utils/date';
 import { HEADER } from 'src/config/componentSizes';
 import { useSearchParams } from 'react-router-dom';
-import { accommodationEventTransform } from '../../hooks/useAccommodationsAndOffers/helpers';
+import { accommodationEventTransform } from '../../utils/useAccommodationsAndOffers';
 import Draggable from 'react-draggable';
 
 export enum ResultsMode {
