+++ src/utils/checkout.ts
@@ -1,7 +1,7 @@
 import { BillingAddress, OrganizerInformation } from '@windingtree/glider-types/dist/win';
 import { OfferCheckoutType } from 'src/store/types';
 import { getTotalRoomCountReducer } from 'src/utils/offers';
-import { getGroupMode } from '../useAccommodationsAndOffers/helpers';
+import { getGroupMode } from './useAccommodationsAndOffers';
 export type BookingModeType = 'group' | 'normal' | undefined;
 
 /**
